Add README.md

+++ src/main/java/com/tmall/service/impl/UserServiceImpl.java
@@ -177,7 +177,7 @@ public class UserServiceImpl implements IUserService {
     // 商城管理后端
 
     public ServerResponse checkAdminRole(User user) {
-        if (user != null && user.getRole().intValue() == Const.Role.ROLE_ADMIN) {
+        if (user != null && user.getRole() == Const.Role.ROLE_ADMIN) {
             return ServerResponse.createBySuccess();
         }
         return ServerResponse.createByError();
